refactor: Make an App Store metadata repository
