+++ packages/app-store/_appRegistry.ts
@@ -1,5 +1,4 @@
 import { appStoreMetadata } from "@calcom/app-store/appStoreMetaData";
-import { getAppFromSlug } from "@calcom/app-store/utils";
 import getInstallCountPerApp from "@calcom/lib/apps/getInstallCountPerApp";
 import { getAllDelegationCredentialsForUser } from "@calcom/lib/delegationCredential/server";
 import type { UserAdminTeams } from "@calcom/lib/server/repository/user";
@@ -7,6 +6,8 @@ import prisma, { safeAppSelect, safeCredentialSelect } from "@calcom/prisma";
 import { userMetadata } from "@calcom/prisma/zod-utils";
 import type { AppFrontendPayload as App } from "@calcom/types/App";
 import type { CredentialFrontendPayload as Credential } from "@calcom/types/Credential";
+
+import { AppStoreMetadataRepository } from "./appStoreMetadataRepository";
 
 export type TDependencyData = {
   name?: string;
@@ -64,6 +65,7 @@ export async function getAppRegistry() {
 
 export async function getAppRegistryWithCredentials(userId: number, userAdminTeams: UserAdminTeams = []) {
   // Get teamIds to grab existing credentials
+  const appStoreMetadataRepository = new AppStoreMetadataRepository();
 
   const dbApps = await prisma.app.findMany({
     where: { enabled: true },
@@ -121,7 +123,7 @@ export async function getAppRegistryWithCredentials(userId: number, userAdminTea
           (dbAppIterator) => dbAppIterator.credentials.length && dbAppIterator.slug === dependency
         );
         // If the app marked as dependency is simply deleted from the codebase, we can have the situation where App is marked installed in DB but we couldn't get the app.
-        const dependencyName = getAppFromSlug(dependency)?.name;
+        const dependencyName = appStoreMetadataRepository.getAppFromSlug(dependency)?.name;
         return { name: dependencyName, installed: dependencyInstalled };
       });
     }
