+++ packages/lib/event-types/getBulkEventTypes.ts
@@ -1,8 +1,9 @@
-import { getAppFromLocationValue } from "@calcom/app-store/utils";
+import { AppStoreMetadataRepository } from "@calcom/app-store/appStoreMetadataRepository";
 import { prisma } from "@calcom/prisma";
 import { eventTypeLocations as eventTypeLocationsSchema } from "@calcom/prisma/zod-utils";
 
 const getBulkEventTypes = async (userId: number) => {
+  const appStoreMetadataRepository = new AppStoreMetadataRepository();
   const eventTypes = await prisma.eventType.findMany({
     where: {
       userId,
@@ -19,7 +20,7 @@ const getBulkEventTypes = async (userId: number) => {
     const locationParsed = eventTypeLocationsSchema.safeParse(eventType.locations);
 
     // some events has null as location for legacy reasons, so this fallbacks to daily video
-    const app = getAppFromLocationValue(
+    const app = appStoreMetadataRepository.getAppFromLocationValue(
       locationParsed.success && locationParsed.data?.[0]?.type
         ? locationParsed.data[0].type
         : "integrations:daily"
