+++ packages/lib/getPaymentAppData.ts
@@ -1,9 +1,9 @@
 import type { z } from "zod";
 
-import { getEventTypeAppData } from "@calcom/app-store/_utils/getEventTypeAppData";
 import type { appDataSchemas } from "@calcom/app-store/apps.schemas.generated";
 import type { appDataSchema, paymentOptionEnum } from "@calcom/app-store/stripepayment/zod";
 import type { EventTypeAppsList } from "@calcom/app-store/utils";
+import { getEventTypeAppData } from "@calcom/app-store/utils";
 import type { BookerEvent } from "@calcom/features/bookings/types";
 import type { EventTypeMetaDataSchema } from "@calcom/prisma/zod-utils";
 import { eventTypeMetaDataSchemaWithTypedApps } from "@calcom/prisma/zod-utils";
