+++ packages/lib/server/getDefaultLocations.ts
@@ -1,6 +1,6 @@
 import getAppKeysFromSlug from "@calcom/app-store/_utils/getAppKeysFromSlug";
+import { AppStoreMetadataRepository } from "@calcom/app-store/appStoreMetadataRepository";
 import { DailyLocationType } from "@calcom/app-store/locations";
-import getApps from "@calcom/app-store/utils";
 import { getUsersCredentials } from "@calcom/lib/server/getUsersCredentials";
 import { userMetadata as userMetadataSchema } from "@calcom/prisma/zod-utils";
 import type { EventTypeLocation } from "@calcom/prisma/zod/custom/eventtype";
@@ -14,14 +14,15 @@ type User = {
 };
 
 export async function getDefaultLocations(user: User): Promise<EventTypeLocation[]> {
+  const appStoreMetadataRepository = new AppStoreMetadataRepository();
   const defaultConferencingData = userMetadataSchema.parse(user.metadata)?.defaultConferencingApp;
 
   if (defaultConferencingData && defaultConferencingData.appSlug !== "daily-video") {
     const credentials = await getUsersCredentials(user);
 
-    const foundApp = getApps(credentials, true).filter(
-      (app) => app.slug === defaultConferencingData.appSlug
-    )[0]; // There is only one possible install here so index [0] is the one we are looking for ;
+    const foundApp = appStoreMetadataRepository
+      .getApps(credentials, true)
+      .filter((app) => app.slug === defaultConferencingData.appSlug)[0]; // There is only one possible install here so index [0] is the one we are looking for ;
     const locationType = foundApp?.locationOption?.value ?? DailyLocationType; // Default to Daily if no location type is found
     return [{ type: locationType, link: defaultConferencingData.appLink }];
   }
