+++ packages/platform/atoms/event-types/hooks/useAtomEventTypeAppIntegration.ts
@@ -1,6 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 
-import type { CredentialDataWithTeamName, LocationOption } from "@calcom/app-store/utils";
+import type { CredentialDataWithTeamName } from "@calcom/app-store/appStoreMetadataRepository";
+import type { LocationOption } from "@calcom/app-store/utils";
 import { SUCCESS_STATUS } from "@calcom/platform-constants";
 import type { ApiResponse } from "@calcom/platform-types";
 import type { App } from "@calcom/types/App";
